+++ test/tests/transition_tests.dart
@@ -1,105 +1,81 @@
 // Copyright (c) 2013, akserg (Sergey Akopkokhyants)
-// https://github.com/akserg/monomer
+// https://github.com/akserg/angular.dart.ui
 // All rights reserved.  Please see the LICENSE.md file.
 
 part of angular.ui.test;
 
 void transitionTests() {
 
-  // TODO: Add tests for IE < 10
   group('Testing transition:', () {
+    
     // Add injector and configure Modules before tests
     setUp(() {
       setUpInjector();
-      module((Module m) => m.install(new AngularUI()));
+      module((Module m) => m.install(new TransitionModule()));
     });
     
     // Remove injector after tests
     tearDown(tearDownInjector);
     
-    /**
-     * Test load Transition
-     */
     test('should load transition', inject((Transition transition) {
       expect(transition, isNot(isEmpty));
     }));
     
-    /**
-     * Test Custorm [Future]
-     */
-    test('returns our custom future', inject((Transition $transition, Scope $rootScope) {
-      var element = new DivElement();
+    test('returns our custom future', inject((Transition transition) {
+      var element = new dom.DivElement();
       
-      var future = $transition.make($rootScope, element, '');
+      var future = transition(element, '').future;
       
       expect(future.then, isNotNull);
       expect(future.catchError, isNotNull);
     }));
     
-    /**
-     * Test CSS if passing a string
-     */
-    test('changes the css if passed a string', inject((Transition $transition, Scope $rootScope) {
-      var element = new DivElement();
+    test('changes the css if passed a string', inject((Transition transition) {
+      var element = new dom.DivElement();
       
-      $transition.make($rootScope, element, 'triggerClass');
-
-      expect(element.classes.contains('triggerClass'), isTrue);
+      transition(element, 'triggerClass').future.then((value) {
+        expect(element.classes.contains('triggerClass'), isTrue);
+      });
     }));
     
-    /**
-     * Test Style if passing an object
-     */
-    test('changes the style if passed an object', inject((Transition $transition, Scope $rootScope) {
-      var element = new DivElement();
+    test('changes the style if passed an object', inject((Transition transition) {
+      var element = new dom.DivElement();
       var triggerStyle = { 'height': '11px' };
       
-      $transition.make($rootScope, element, triggerStyle);
-
-      expect(element.style.getPropertyValue('height'), equals('11px'));
-    }));
-    
-    /**
-     * Test calls the function.
-     */
-    test('calls the function if passed', inject((Transition $transition, Scope $rootScope) {
-      var element = new DivElement();
-
-      $transition.make($rootScope, element, (Element e){
-        e.classes.add("test");
+      transition(element, triggerStyle).future.then((value) {
+        expect(element.style.getPropertyValue('height'), equals('11px'));
       });
-
-      expect(element.classes.contains("test"), isTrue);
     }));
     
-    /**
-     * Test transitionEndEventName bust be equals null
-     */
-    test('should be undefined', inject((Transition $transition, Scope $rootScope) {
-      expect($transition.transitionEndEventName == null, isTrue);
+    test('calls the function if passed', inject((Transition transition) {
+      var element = new dom.DivElement();
+
+      transition(element, (dom.Element e){
+        e.classes.add("test");
+      }).future.then((value) {
+        expect(element.classes.contains("test"), isTrue);
+      });
     }));
     
-    /**
-     * Test a transitionEnd handler does not bind to element
-     */
-    test('does not bind a transitionEnd handler to the element', inject((Transition $transition, Scope $rootScope) {
-      var element = new DivElement();
-
-      $transition.make($rootScope, element, '');
-
-      expect($transition.transitionEndEventName, isNull);
+    test('should be undefined', inject((Transition transition) {
+      expect(transition.transitionEndEventName == null, isTrue);
     }));
     
-    /**
-     * Test future
-     */
-    test('resolves the future', inject((Transition $transition, Scope $rootScope) {
-      var element = new DivElement();
-      var transitionEndHandler = (Element e) {
+    test('does not bind a transitionEnd handler to the element', inject((Transition transition) {
+      var element = new dom.DivElement();
+
+      transition(element, '').future.then((value) {
+        expect(transition.transitionEndEventName, isNull);
+      });
+    }));
+    
+    test('resolves the future', inject((Transition transition) {
+      var element = new dom.DivElement();
+      var transitionEndHandler = (dom.Element e) {
         e.classes.add("test");
       };
       
-      Future future = $transition.make($rootScope, element, '');
+      var future = transition(element, '').future;
       future.then(transitionEndHandler).then((_){
         expect(element.classes.contains("test"), isTrue);
       });