+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 Angular UI Changes
 ===============
 
-### Release 0.1, In Progress
+## Release 0.2.0, In Progress...
 
-- Created Transition.
+- Created Collapse module. Provides a simple way to hide and show an element with a css transition.
+
+## Release 0.1.0, Released
+
+### 2014/01/14
+- Created Buttons module. Module contains two directives that can make a group of buttons to behave like a set of checkboxes or radio buttons.
+
+### 2014/01/11
+- Created Position and Transition modules.